requirements.txt
# Project dependencies.
gradio
transformers
torch
pdfplumber
python-docx
SpeechRecognition
pydub
-e .